Update TilePlace.java
# TilePlace Java Program

## Overview

The `TilePlace` Java program provides a simple implementation for calculating the number of unique paths to move from the top-left corner of a grid to the bottom-right corner. The grid consists of tiles, and the program allows movement either horizontally or vertically.

## Functionality

The core functionality is implemented in the `paths` method. Given the dimensions of the grid (`n` rows and `m` columns), the method recursively calculates the number of paths to reach the destination.

- If the number of rows (`n`) is equal to the number of columns (`m`), there are two possible paths: horizontal and vertical.
- If the number of rows is less than the number of columns, only a vertical path is allowed.
- Otherwise, the method recursively calculates both horizontal and vertical paths, summing up the results.

## Usage

The `main` method is used to demonstrate the functionality of the `paths` method. You can customize the values of `n` and `m` to observe the total number of paths for different grid dimensions.

```java
public static void main(String[] args) {
    int n = 4, m = 2;
    int totalPaths = paths(n, m);
    System.out.println("Total Paths: " + totalPaths);
}